+++ src/Router.php
@@ -57,10 +57,6 @@ class Router
 	 */
 	public function run(): never
 	{
-		// Set a constant to check if current request is XHR
-		if (!defined('IS_XHR')) {
-			define('IS_XHR', HttpUtils::getHeader('X-Requested-With') == 'XMLHttpRequest');
-		}
 		$requestMethod = $_SERVER['REQUEST_METHOD'];
 
 		try {